requirements.txt
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
ttkthemes==3.2.2
